+++ lib/ui/shared/sizes.dart
@@ -1,0 +1,7 @@
+class Sizes {
+  static const double small = 4;
+  static const double medium = 8;
+  static const double large = 16;
+  static const double extraLarge = 32;
+  static const double extraExtraLarge = 64;
+}
