+++ lib/ui/widgets/username.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:portfolio/core/utils/constants.dart';
+import 'package:portfolio/ui/shared/sizes.dart';
 import 'package:portfolio/ui/widgets/text/body_medium_text.dart';
 
 class Username extends StatelessWidget {
@@ -7,7 +8,7 @@ class Username extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return const Padding(
-      padding: EdgeInsets.all(16),
+      padding: EdgeInsets.all(Sizes.large),
       child: BodyMediumText(Constants.username),
     );
   }
